Depression Chatbot/requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.2.0
apparmor==3.1.7
argcomplete==2.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
autocommand==2.2.2
Babel==2.12.1
beautifulsoup4==4.12.2
bleach==6.1.0
btrfsutil==6.7
build==1.0.3
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
dbus-python==1.3.2
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
docutils==0.20.1
evdev==1.6.1
executing==2.0.0
fastjsonschema==2.19.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.48.1
future==0.18.3
gast==0.5.4
gbinder-python==1.1.2
gmpy2==2.1.5
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.1
h11==0.14.0
h5py==3.10.0
html5lib==1.1
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib_metadata==5.1.0
inflect==7.0.0
installer==0.7.0
ipykernel==6.29.2
ipython==8.21.0
jaraco.context==4.3.0
jaraco.functools==4.0.0
jaraco.text==3.12.0
jedi==0.19.1
Jinja2==3.1.3
json5==0.9.14
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyterlab==4.1.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==2.12.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
LibAppArmor==3.1.7
libinput-gestures-qt==0.3
lxml==4.9.2
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mercurial==6.6.3
mistune==2.0.5
ml_dtypes==0.3.2
moddb==0.10.0
more-itertools==10.1.0
mpmath==1.3.0
nbclient==0.9.0
nbconvert==7.16.0
nbformat==5.9.2
nest_asyncio==1.5.9
networkx==3.2.1
nftables==0.1
nltk==3.8.1
notebook==7.1.0
notebook_shim==0.2.3
numpy==1.26.3
oauthlib==3.2.2
opt-einsum==3.3.0
ordered-set==4.1.0
overrides==7.7.0
packaging==23.2
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
ply==3.11
pooch==1.8.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pwquality==1.4.5
pyasn1==0.5.0
pyasn1_modules==0.3.0
pybind11==2.11.1
pycairo==1.25.1
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydot==1.4.2
Pygments==2.17.2
PyGObject==3.46.0
pyparsing==3.1.1
pyproject_hooks==1.0.0
PyQt5==5.15.10
PyQt5-sip==12.13.0
pyrate-limiter==3.2.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
QTermWidget==1.4.0
referencing==0.33.0
Reflector==2023.6.28.0.36.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
scipy==1.12.0
Send2Trash==1.8.2
six==1.16.0
smbus==1.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
TBB==0.2
tensorboard==2.15.1
tensorboard-plugin-wit==1.8.1
tensorboard_data_server==0.8.0a0
tensorflow==2.15.0
tensorflow-estimator==2.15.0
termcolor==2.4.0
tinycss2==1.2.1
tomli==2.0.1
toolz==0.12.1
torch==2.2.0
tornado==6.3.2
tqdm==4.66.2
traitlets==5.14.1
trove-classifiers==2024.2.2
typing_extensions==4.9.0
uc-micro-py==1.0.3
ufw==0.36.2
urllib3==1.26.18
validate-pyproject==0.13.post1.dev0+gb752273.d20230520
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==2.3.8
wrapt==1.14.1
zipp==3.17.0
